added disappearing messages

+++ base/views.py
@@ -3,6 +3,7 @@ from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 
 from .models import Room, Topic
 from .forms import RoomForm
@@ -53,7 +54,7 @@ def room(request, pk):
     context = {"room": room}
     return render(request, "base/room.html", context=context)
 
-
+@login_required(login_url='/login')
 def create_room(request):
     form = RoomForm()
 
@@ -61,34 +62,46 @@ def create_room(request):
         form = RoomForm(request.POST)
         if form.is_valid():
             form.save()
+            messages.success(request, 'Room creation successful!!!.')
             return redirect("home")
     context = {"form": form}
+    
     return render(request, "base/room_form.html", context)
 
-
+@login_required(login_url='/login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
-
     form = RoomForm(instance=room)
 
+    if request.user != room.host:
+        messages.error(request, 'You are not allowed to update or delete rooms that you do not host.')
+        return redirect('home')
+    
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
             form.save()
+            messages.success(request, 'Room update successful!!!.')
             return redirect("home")
 
     context = {"form": form}
+    
     return render(request, "base/room_form.html", context)
 
-
+@login_required(login_url='/login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
 
+    if request.user != room.host:
+        messages.error(request, 'You are not allowed to update or delete rooms that you do not host.')
+        return redirect('home')
+    
     if request.method == "POST":
         room.delete()
+        messages.success(request, 'Room deletion successful!!!.')
         return redirect("home")
     context = {
         "obj": room,
     }
-
+    
     return render(request, "base/delete.html", context)
